pylintrc: re-enable too-many-lines, part 5
To get under the 1000 line limit, move host_helpers
test classes into unique files under
tests/unit/host_helpers/.
